GameWindow
Looked for ways to improve slow redraw issue. I think the real problem is that TableLayoutPanel is an old control and notoriously slow (found lots of hits googling "TableLayoutPanel slow".
Injecting new IImageProvider dependency into this constructor as a means of getting images.

CachedImageProvider

Implementation of IImageProvider that simplifies retrieval of images, and provides caching via a dictionary so that images only have to be loaded from disk once. It doesn't make much of a difference here, but this provides a pattern for possible future use when larger/more file reads start affecting application performance.

Field

Added a static Create method with all the board size logic. Seemed misplaced in GameWindow. Also moved GameSize enum to this class so we could just rely on it other than reading from the initial Menu Button text.